Breakout metrics into `fuel-metrics` (#589)
* moved metrics to their own module

* rearrange tomls for linter

* remove unused prometheus dep from fuel-core
