+++ fuel-core/src/state/rocks_db.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "prometheus")]
-use crate::service::metrics::prometheus_metrics::DATABASE_METRICS;
 use crate::state::KVItem;
 use crate::{
     database::{
@@ -12,6 +10,8 @@ use crate::{
         WriteOperation,
     },
 };
+#[cfg(feature = "metrics")]
+use fuel_metrics::core_metrics::DATABASE_METRICS;
 use rocksdb::{
     BoundColumnFamily, ColumnFamilyDescriptor, DBCompressionType, DBWithThreadMode, IteratorMode,
     MultiThreaded, Options, ReadOptions, SliceTransform, WriteBatch,
@@ -106,13 +106,13 @@ impl RocksDb {
 
 impl KeyValueStore for RocksDb {
     fn get(&self, key: &[u8], column: ColumnId) -> crate::state::Result<Option<Vec<u8>>> {
-        #[cfg(feature = "prometheus")]
+        #[cfg(feature = "metrics")]
         DATABASE_METRICS.read_meter.inc();
         let value = self
             .db
             .get_cf(&self.cf(column), key)
             .map_err(|e| Error::DatabaseError(Box::new(e)));
-        #[cfg(feature = "prometheus")]
+        #[cfg(feature = "metrics")]
         {
             if value.is_ok() && value.as_ref().unwrap().is_some() {
                 let value_as_vec = value.as_ref().cloned().unwrap().unwrap();
@@ -130,7 +130,7 @@ impl KeyValueStore for RocksDb {
         column: ColumnId,
         value: Vec<u8>,
     ) -> crate::state::Result<Option<Vec<u8>>> {
-        #[cfg(feature = "prometheus")]
+        #[cfg(feature = "metrics")]
         {
             DATABASE_METRICS.write_meter.inc();
             DATABASE_METRICS
@@ -199,7 +199,7 @@ impl KeyValueStore for RocksDb {
                 item.map(|(key, value)| {
                     let value_as_vec = value.to_vec();
                     let key_as_vec = key.to_vec();
-                    #[cfg(feature = "prometheus")]
+                    #[cfg(feature = "metrics")]
                     {
                         DATABASE_METRICS.read_meter.inc();
                         DATABASE_METRICS
@@ -241,7 +241,7 @@ impl BatchOperations for RocksDb {
                 }
             }
         }
-        #[cfg(feature = "prometheus")]
+        #[cfg(feature = "metrics")]
         {
             DATABASE_METRICS.write_meter.inc();
             DATABASE_METRICS
