+++ fuel-metrics/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "fuel-metrics"
+version = "0.10.1"
+authors = ["Fuel Labs <contact@fuel.sh>"]
+categories = ["cryptography::cryptocurrencies"]
+edition = "2021"
+homepage = "https://fuel.network/"
+keywords = ["blockchain", "cryptocurrencies", "fuel-vm", "vm"]
+license = "BUSL-1.1"
+repository = "https://github.com/FuelLabs/fuel-core"
+description = "Fuel metrics"
+
+[dependencies]
+axum = "0.4"
+hyper = "0.14"
+lazy_static = "1.4"
+prometheus = "0.13"
+
